Tests/LibWeb/Text/expected/UIEvents/wheel-event-legacy-deltas.txt

wheelDeltaX: 0
wheelDeltaY: -100
wheelDelta: -100
scrollY: 100
